% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{adm_layout}
\alias{adm_layout}
\title{Actuarial Data Master Helper Datasets}
\format{
A data frame with 1976 rows and 17 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(adm_layout)
}
\description{
A combined dataset for adm_layout
}
\keyword{datasets}
