% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{AdmLayout}
\alias{AdmLayout}
\title{Actuarial Data Master Helper Datasets}
\format{
A data frame with 17755 rows and 18 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(AdmLayout)
}
\description{
A combined dataset for AdmLayout
}
\keyword{datasets}
