% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{agr_commodity}
\alias{agr_commodity}
\title{agr_commodity}
\format{
A data frame with 798 rows and 13 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(agr_commodity)
}
\description{
A combined dataset for agr_commodity
}
\keyword{datasets}
