% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{AgrCommodity}
\alias{AgrCommodity}
\title{AgrCommodity}
\format{
A data frame with 6170 rows and 15 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(AgrCommodity)
}
\description{
A combined dataset for AgrCommodity
}
\keyword{datasets}
