% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{area_data_source}
\alias{area_data_source}
\title{area_data_source}
\format{
A data frame with 985 rows and 21 columns covering 2017-2025.

A data frame with 985 rows and 21 columns covering 2017-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(area_data_source)

data(area_data_source)
}
\description{
A combined dataset for area_data_source

A combined dataset for area_data_source
}
\keyword{datasets}
