% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{AreaDataSource}
\alias{AreaDataSource}
\title{AreaDataSource}
\format{
A data frame with 985 rows and 21 columns covering 2017-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(AreaDataSource)
}
\description{
A combined dataset for AreaDataSource
}
\keyword{datasets}
