% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Beta}
\alias{Beta}
\title{Beta}
\format{
A data frame with 1887000 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Beta)
}
\description{
A combined dataset for Beta
}
\keyword{datasets}
