% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{class}
\alias{class}
\title{class}
\format{
A data frame with 51 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(class)
}
\description{
A combined dataset for class
}
\keyword{datasets}
