% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_file_name}
\alias{clean_file_name}
\title{Clean the file name}
\usage{
clean_file_name(file_name, file_type_out = "rds")
}
\arguments{
\item{file_name}{the file path of the file name to clean}

\item{file_type_out}{the type of file (currently only supports rds)}
}
\value{
a version of the file name that is cleaned (i.e. snake case, .rds suffix, no extraneous information)
}
\description{
Clean the file name
}
\examples{
\dontrun{clean_file_name("./data-raw/2012_A01100_YieldAndTyield_YTD.txt")}
}
