% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{commodity}
\alias{commodity}
\title{commodity}
\format{
A data frame with 267 rows and 11 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(commodity)
}
\description{
A combined dataset for commodity
}
\keyword{datasets}
