% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{commodity_type}
\alias{commodity_type}
\title{commodity_type}
\format{
A data frame with 682 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(commodity_type)
}
\description{
A combined dataset for commodity_type
}
\keyword{datasets}
