% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{CommodityType}
\alias{CommodityType}
\title{CommodityType}
\format{
A data frame with 5593 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(CommodityType)
}
\description{
A combined dataset for CommodityType
}
\keyword{datasets}
