% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_adm.R, R/helpers.R
\name{compress_adm}
\alias{compress_adm}
\title{Compress ADM files by maintaining only the necessary level of granularity}
\usage{
compress_adm(table_code, df, dir)

compress_adm(table_code, df, dir)
}
\arguments{
\item{table_code}{An adm record code to identify the target record.}

\item{df}{a data frame corresponding to the data represented by the table_code}

\item{dir}{Path to where downloaded adm files are stored.}
}
\value{
A data frame of aggregated parameter values.

A data frame of aggregated parameter values.
}
\description{
Compress ADM files by maintaining only the necessary level of granularity

Compress ADM files by maintaining only the necessary level of granularity
}
