% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{contiguous_county}
\alias{contiguous_county}
\title{contiguous_county}
\format{
A data frame with 104103 rows and 11 columns covering 2020-2025.

A data frame with 104103 rows and 11 columns covering 2020-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(contiguous_county)

data(contiguous_county)
}
\description{
A combined dataset for contiguous_county

A combined dataset for contiguous_county
}
\keyword{datasets}
