% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ContiguousCounty}
\alias{ContiguousCounty}
\title{ContiguousCounty}
\format{
A data frame with 104103 rows and 11 columns covering 2020-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(ContiguousCounty)
}
\description{
A combined dataset for ContiguousCounty
}
\keyword{datasets}
