% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{county}
\alias{county}
\title{county}
\format{
A data frame with 6375 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(county)
}
\description{
A combined dataset for county
}
\keyword{datasets}
