% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{cropping_practice}
\alias{cropping_practice}
\title{cropping_practice}
\format{
A data frame with 144 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(cropping_practice)
}
\description{
A combined dataset for cropping_practice
}
\keyword{datasets}
