% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{CroppingPractice}
\alias{CroppingPractice}
\title{CroppingPractice}
\format{
A data frame with 1263 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(CroppingPractice)
}
\description{
A combined dataset for CroppingPractice
}
\keyword{datasets}
