% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{document_structure}
\alias{document_structure}
\title{document_structure}
\format{
A data frame with 19 rows and 11 columns covering 2013-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(document_structure)
}
\description{
A combined dataset for document_structure
}
\keyword{datasets}
