% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{DocumentStructure}
\alias{DocumentStructure}
\title{DocumentStructure}
\format{
A data frame with 274 rows and 11 columns covering 2013-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(DocumentStructure)
}
\description{
A combined dataset for DocumentStructure
}
\keyword{datasets}
