% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{endorsement_length}
\alias{endorsement_length}
\title{endorsement_length}
\format{
A data frame with 13 rows and 8 columns covering 2013-2025.

A data frame with 13 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(endorsement_length)

data(endorsement_length)
}
\description{
A combined dataset for endorsement_length

A combined dataset for endorsement_length
}
\keyword{datasets}
