% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{EndorsementLength}
\alias{EndorsementLength}
\title{EndorsementLength}
\format{
A data frame with 13 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(EndorsementLength)
}
\description{
A combined dataset for EndorsementLength
}
\keyword{datasets}
