% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_FORCE_NUMERIC_KEYS}
\alias{FCIP_FORCE_NUMERIC_KEYS}
\title{Column names to coerce to numeric}
\format{
An object of class \code{character} of length 9.
}
\usage{
FCIP_FORCE_NUMERIC_KEYS
}
\description{
A character vector of column names that should be converted from character
to numeric during data ingestion and cleaning.
}
\examples{
\dontrun{
# View default keys
FCIP_FORCE_NUMERIC_KEYS
}

}
\keyword{datasets}
