% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_INSURANCE_ELECTION}
\alias{FCIP_INSURANCE_ELECTION}
\title{Insurance election identifier fields}
\format{
A \code{character} vector of field names:
}
\usage{
FCIP_INSURANCE_ELECTION
}
\value{
A \code{character} vector of field names used to specify insurance elections.
}
\description{
A character vector of column names that define an insurance election within
the Federal Crop Insurance Program (FCIP). Each field corresponds to an
attribute of a policy election.
}
\examples{
\dontrun{
# Default election fields
FCIP_INSURANCE_ELECTION
}
}
\keyword{datasets}
