% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_INSURANCE_ELECTION_RCODED}
\alias{FCIP_INSURANCE_ELECTION_RCODED}
\title{Insurance election identifier fields (recoded)}
\format{
A \code{character} vector of field names:
}
\usage{
FCIP_INSURANCE_ELECTION_RCODED
}
\description{
A character vector of recoded column names that specify an insurance
election within the Federal Crop Insurance Program (FCIP). These fields
correspond to recoded versions of the original election attributes.
}
\examples{
\dontrun{
# View the default recoded election fields
FCIP_INSURANCE_ELECTION_RCODED

}
}
\keyword{datasets}
