% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_INSURANCE_POOL}
\alias{FCIP_INSURANCE_POOL}
\title{Insurance pool identifier fields}
\format{
A \code{character} vector of field names.
}
\usage{
FCIP_INSURANCE_POOL
}
\value{
A \code{character} vector specifying the columns used to define
each FCIP insurance pool.
}
\description{
A character vector of column names that together define a unique insurance
pool in the Federal Crop Insurance Program (FCIP).
}
\examples{
\dontrun{
# Default insurance pool fields
FCIP_INSURANCE_POOL
}
}
\keyword{datasets}
