% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{hip_rate}
\alias{hip_rate}
\title{hip_rate}
\format{
A data frame with 50896 rows and 13 columns covering 2020-2025.

A data frame with 50896 rows and 13 columns covering 2020-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(hip_rate)

data(hip_rate)
}
\description{
A combined dataset for hip_rate

A combined dataset for hip_rate
}
\keyword{datasets}
