% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{hip_triggering_event}
\alias{hip_triggering_event}
\title{hip_triggering_event}
\format{
A data frame with 55 rows and 8 columns covering Inf--Inf.

A data frame with 55 rows and 8 columns covering Inf--Inf.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(hip_triggering_event)

data(hip_triggering_event)
}
\description{
A combined dataset for hip_triggering_event

A combined dataset for hip_triggering_event
}
\keyword{datasets}
