% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{HipRate}
\alias{HipRate}
\title{HipRate}
\format{
A data frame with 51013 rows and 13 columns covering 2020-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(HipRate)
}
\description{
A combined dataset for HipRate
}
\keyword{datasets}
