% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{HipTriggeringEvent}
\alias{HipTriggeringEvent}
\title{HipTriggeringEvent}
\format{
A data frame with 57 rows and 8 columns covering Inf--Inf.
}
\source{
Actuarial Data Master
}
\usage{
data(HipTriggeringEvent)
}
\description{
A combined dataset for HipTriggeringEvent
}
\keyword{datasets}
