% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{insurance_option}
\alias{insurance_option}
\title{insurance_option}
\format{
A data frame with 702 rows and 8 columns covering 2013-2025.

A data frame with 702 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(insurance_option)

data(insurance_option)
}
\description{
A combined dataset for insurance_option

A combined dataset for insurance_option
}
\keyword{datasets}
