% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{insurance_plan}
\alias{insurance_plan}
\title{insurance_plan}
\format{
A data frame with 412 rows and 10 columns covering 2011-2025.

A data frame with 412 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(insurance_plan)

data(insurance_plan)
}
\description{
A combined dataset for insurance_plan

A combined dataset for insurance_plan
}
\keyword{datasets}
