% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{InsuranceOption}
\alias{InsuranceOption}
\title{InsuranceOption}
\format{
A data frame with 702 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(InsuranceOption)
}
\description{
A combined dataset for InsuranceOption
}
\keyword{datasets}
