% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{InsurancePlan}
\alias{InsurancePlan}
\title{InsurancePlan}
\format{
A data frame with 412 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(InsurancePlan)
}
\description{
A combined dataset for InsurancePlan
}
\keyword{datasets}
