% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{intended_use}
\alias{intended_use}
\title{intended_use}
\format{
A data frame with 186 rows and 9 columns covering 2011-2025.

A data frame with 186 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(intended_use)

data(intended_use)
}
\description{
A combined dataset for intended_use

A combined dataset for intended_use
}
\keyword{datasets}
