% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{IntendedUse}
\alias{IntendedUse}
\title{IntendedUse}
\format{
A data frame with 186 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(IntendedUse)
}
\description{
A combined dataset for IntendedUse
}
\keyword{datasets}
