% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{interval}
\alias{interval}
\title{interval}
\format{
A data frame with 90 rows and 11 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(interval)
}
\description{
A combined dataset for interval
}
\keyword{datasets}
