% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{irrigation_practice}
\alias{irrigation_practice}
\title{irrigation_practice}
\format{
A data frame with 6 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(irrigation_practice)
}
\description{
A combined dataset for irrigation_practice
}
\keyword{datasets}
