% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{IrrigationPractice}
\alias{IrrigationPractice}
\title{IrrigationPractice}
\format{
A data frame with 45 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(IrrigationPractice)
}
\description{
A combined dataset for IrrigationPractice
}
\keyword{datasets}
