% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_numeric_convertible}
\alias{is_numeric_convertible}
\title{Test if a column can be converted to numeric without data loss}
\usage{
is_numeric_convertible(x, col_name = NULL)
}
\arguments{
\item{x}{A vector to test for numeric convertibility}
}
\value{
Logical. TRUE if all non-NA values can be converted to numeric without loss
}
\description{
Test if a column can be converted to numeric without data loss
}
\keyword{internal}
