% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{organic_practice}
\alias{organic_practice}
\title{organic_practice}
\format{
A data frame with 6 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(organic_practice)
}
\description{
A combined dataset for organic_practice
}
\keyword{datasets}
