% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{OrganicPractice}
\alias{OrganicPractice}
\title{OrganicPractice}
\format{
A data frame with 45 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(OrganicPractice)
}
\description{
A combined dataset for OrganicPractice
}
\keyword{datasets}
