% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{PaceDate}
\alias{PaceDate}
\title{PaceDate}
\format{
A data frame with 133801 rows and 12 columns covering 2022-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(PaceDate)
}
\description{
A combined dataset for PaceDate
}
\keyword{datasets}
