% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{practice}
\alias{practice}
\title{practice}
\format{
A data frame with 17660 rows and 10 columns covering 2011-2025.

A data frame with 17660 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(practice)

data(practice)
}
\description{
A combined dataset for practice

A combined dataset for practice
}
\keyword{datasets}
