% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_county_yield_history.R
\name{process_single_archive}
\alias{process_single_archive}
\title{Process data for a single year}
\usage{
process_single_archive(year)
}
\arguments{
\item{year}{\link{numeric} Year to process data for}
}
\value{
\link[data.table]{data.table} with aggregated yield data, or NULL if processing fails
}
\description{
Loads and processes insurance offer and yield trend data for a single year,
aggregating yield data by mean within FCIP insurance pools, commodity years, and yield years.
}
