% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ProductionArea}
\alias{ProductionArea}
\title{ProductionArea}
\format{
A data frame with 1775968 rows and 11 columns covering 2017-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(ProductionArea)
}
\description{
A combined dataset for ProductionArea
}
\keyword{datasets}
