% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{quality}
\alias{quality}
\title{quality}
\format{
A data frame with 8006 rows and 12 columns covering 2011-2025.

A data frame with 8006 rows and 12 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(quality)

data(quality)
}
\description{
A combined dataset for quality

A combined dataset for quality
}
\keyword{datasets}
