% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{rate_method}
\alias{rate_method}
\title{rate_method}
\format{
A data frame with 52 rows and 8 columns covering 2013-2025.

A data frame with 52 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(rate_method)

data(rate_method)
}
\description{
A combined dataset for rate_method

A combined dataset for rate_method
}
\keyword{datasets}
