% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{RateMethod}
\alias{RateMethod}
\title{RateMethod}
\format{
A data frame with 52 rows and 8 columns covering 2013-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(RateMethod)
}
\description{
A combined dataset for RateMethod
}
\keyword{datasets}
