% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_factor_metadata}
\alias{save_factor_metadata}
\title{Save factor metadata to package data folder}
\usage{
save_factor_metadata(metadata_key, factor_metadata)
}
\arguments{
\item{metadata_key}{Character. Unique key for this dataset}

\item{factor_metadata}{List. Factor level mappings}
}
\description{
Save factor metadata to package data folder
}
\keyword{internal}
