% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{spoi_group}
\alias{spoi_group}
\title{spoi_group}
\format{
A data frame with 46 rows and 12 columns covering 2013-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(spoi_group)
}
\description{
A combined dataset for spoi_group
}
\keyword{datasets}
