% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{SpoiGroup}
\alias{SpoiGroup}
\title{SpoiGroup}
\format{
A data frame with 1494 rows and 12 columns covering 2013-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(SpoiGroup)
}
\description{
A combined dataset for SpoiGroup
}
\keyword{datasets}
