% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{State}
\alias{State}
\title{State}
\format{
A data frame with 750 rows and 11 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(State)
}
\description{
A combined dataset for State
}
\keyword{datasets}
