% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Statement}
\alias{Statement}
\title{Statement}
\format{
A data frame with 776 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Statement)
}
\description{
A combined dataset for Statement
}
\keyword{datasets}
