% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{sub_class}
\alias{sub_class}
\title{sub_class}
\format{
A data frame with 8 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(sub_class)
}
\description{
A combined dataset for sub_class
}
\keyword{datasets}
