% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{SubClass}
\alias{SubClass}
\title{SubClass}
\format{
A data frame with 80 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(SubClass)
}
\description{
A combined dataset for SubClass
}
\keyword{datasets}
