% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{subsidy_percent}
\alias{subsidy_percent}
\title{subsidy_percent}
\format{
A data frame with 670 rows and 16 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(subsidy_percent)
}
\description{
A combined dataset for subsidy_percent
}
\keyword{datasets}
