% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{SubsidyPercent}
\alias{SubsidyPercent}
\title{SubsidyPercent}
\format{
A data frame with 9543 rows and 22 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(SubsidyPercent)
}
\description{
A combined dataset for SubsidyPercent
}
\keyword{datasets}
