% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{type}
\alias{type}
\title{type}
\format{
A data frame with 1026 rows and 10 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(type)
}
\description{
A combined dataset for type
}
\keyword{datasets}
