% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{unit_discount}
\alias{unit_discount}
\title{unit_discount}
\format{
A data frame with 9992 rows and 28 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(unit_discount)
}
\description{
A combined dataset for unit_discount
}
\keyword{datasets}
