% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{adm_layout}
\alias{adm_layout}
\title{Actuarial Data Master Helper Datasets}
\format{
A data frame with 17755 rows and 18 columns covering 2011-2025.

A data frame with 17755 rows and 18 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(adm_layout)

data(adm_layout)
}
\description{
A combined dataset for adm_layout

A combined dataset for adm_layout
}
\keyword{datasets}
