% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{agr_commodity}
\alias{agr_commodity}
\title{agr_commodity}
\format{
A data frame with 6170 rows and 15 columns covering 2011-2025.

A data frame with 6170 rows and 15 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(agr_commodity)

data(agr_commodity)
}
\description{
A combined dataset for agr_commodity

A combined dataset for agr_commodity
}
\keyword{datasets}
