% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{beta}
\alias{beta}
\title{beta}
\format{
A data frame with 242500 rows and 10 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(beta)
}
\description{
A combined dataset for beta
}
\keyword{datasets}
