% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Class}
\alias{Class}
\title{Class}
\format{
A data frame with 413 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Class)
}
\description{
A combined dataset for Class
}
\keyword{datasets}
