% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{class}
\alias{class}
\title{class}
\format{
A data frame with 413 rows and 9 columns covering 2011-2025.

A data frame with 413 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(class)

data(class)
}
\description{
A combined dataset for class

A combined dataset for class
}
\keyword{datasets}
