% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Commodity}
\alias{Commodity}
\title{Commodity}
\format{
A data frame with 2015 rows and 12 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Commodity)
}
\description{
A combined dataset for Commodity
}
\keyword{datasets}
