% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{commodity}
\alias{commodity}
\title{commodity}
\format{
A data frame with 2015 rows and 12 columns covering 2011-2025.

A data frame with 2015 rows and 12 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(commodity)

data(commodity)
}
\description{
A combined dataset for commodity

A combined dataset for commodity
}
\keyword{datasets}
