% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{commodity_type}
\alias{commodity_type}
\title{commodity_type}
\format{
A data frame with 5593 rows and 9 columns covering 2011-2025.

A data frame with 5593 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(commodity_type)

data(commodity_type)
}
\description{
A combined dataset for commodity_type

A combined dataset for commodity_type
}
\keyword{datasets}
