% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{county}
\alias{county}
\title{county}
\format{
A data frame with 47854 rows and 9 columns covering 2011-2025.

A data frame with 47854 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(county)

data(county)
}
\description{
A combined dataset for county

A combined dataset for county
}
\keyword{datasets}
