% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{County}
\alias{County}
\title{County}
\format{
A data frame with 47854 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(County)
}
\description{
A combined dataset for County
}
\keyword{datasets}
