% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{cropping_practice}
\alias{cropping_practice}
\title{cropping_practice}
\format{
A data frame with 1263 rows and 9 columns covering 2011-2025.

A data frame with 1263 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(cropping_practice)

data(cropping_practice)
}
\description{
A combined dataset for cropping_practice

A combined dataset for cropping_practice
}
\keyword{datasets}
