% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{document_structure}
\alias{document_structure}
\title{document_structure}
\format{
A data frame with 274 rows and 11 columns covering 2013-2025.

A data frame with 274 rows and 11 columns covering 2013-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(document_structure)

data(document_structure)
}
\description{
A combined dataset for document_structure

A combined dataset for document_structure
}
\keyword{datasets}
