% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_adm}
\alias{download_adm}
\title{Download the adm files for a given year}
\usage{
download_adm(
  year = 2012,
  adm_url = "https://pubfs-rma.fpac.usda.gov/pub/References/actuarial_data_master/",
  dir = "./data-raw"
)
}
\arguments{
\item{year}{the year of the actuarial data master to download}

\item{adm_url}{the url where the adm FTP site is}

\item{dir}{the directory to save the files to}
}
\value{
the data and layout files for the given year
}
\description{
Download the adm files for a given year
}
\examples{
\dontrun{download_adm(year = 2012)}
}
