% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_adm}
\alias{download_adm}
\title{Download the adm files for a given year}
\usage{
download_adm(
  years = 2012,
  adm_url = "https://pubfs-rma.fpac.usda.gov/pub/References/actuarial_data_master/",
  dir = "./data-raw",
  helpers_only = TRUE,
  keep_source_files = FALSE
)
}
\arguments{
\item{years}{the years of the actuarial data master to download}

\item{adm_url}{the url where the adm FTP site is}

\item{dir}{the directory to save the files to}

\item{helpers_only}{if TRUE, only keeps the helper files (i.e. files smaller than 1 mb)}

\item{keep_source_files}{if TRUE, keeps the original zip files in the year directory. If FALSE, they will be deleted.}
}
\value{
the data and layout files for the given year
}
\description{
This function downloads the raw adm files, applies some minor cleaning operations,
and converts them to .rds format for better compression.
}
\examples{
\dontrun{download_adm(year = 2012)}
}
