% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{endorsement_length}
\alias{endorsement_length}
\title{endorsement_length}
\format{
A data frame with 1 rows and 8 columns covering 2013-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(endorsement_length)
}
\description{
A combined dataset for endorsement_length
}
\keyword{datasets}
