% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{insurance_plan}
\alias{insurance_plan}
\title{insurance_plan}
\format{
A data frame with 40 rows and 10 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(insurance_plan)
}
\description{
A combined dataset for insurance_plan
}
\keyword{datasets}
