% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{intended_use}
\alias{intended_use}
\title{intended_use}
\format{
A data frame with 22 rows and 9 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(intended_use)
}
\description{
A combined dataset for intended_use
}
\keyword{datasets}
