% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Interval}
\alias{Interval}
\title{Interval}
\format{
A data frame with 872 rows and 11 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Interval)
}
\description{
A combined dataset for Interval
}
\keyword{datasets}
