% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{interval}
\alias{interval}
\title{interval}
\format{
A data frame with 872 rows and 11 columns covering 2011-2025.

A data frame with 872 rows and 11 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(interval)

data(interval)
}
\description{
A combined dataset for interval

A combined dataset for interval
}
\keyword{datasets}
