% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{irrigation_practice}
\alias{irrigation_practice}
\title{irrigation_practice}
\format{
A data frame with 45 rows and 9 columns covering 2011-2025.

A data frame with 45 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(irrigation_practice)

data(irrigation_practice)
}
\description{
A combined dataset for irrigation_practice

A combined dataset for irrigation_practice
}
\keyword{datasets}
