% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{organic_practice}
\alias{organic_practice}
\title{organic_practice}
\format{
A data frame with 45 rows and 9 columns covering 2011-2025.

A data frame with 45 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(organic_practice)

data(organic_practice)
}
\description{
A combined dataset for organic_practice

A combined dataset for organic_practice
}
\keyword{datasets}
