% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Practice}
\alias{Practice}
\title{Practice}
\format{
A data frame with 17660 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Practice)
}
\description{
A combined dataset for Practice
}
\keyword{datasets}
