% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{practice}
\alias{practice}
\title{practice}
\format{
A data frame with 2039 rows and 10 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(practice)
}
\description{
A combined dataset for practice
}
\keyword{datasets}
