% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Quality}
\alias{Quality}
\title{Quality}
\format{
A data frame with 8006 rows and 12 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Quality)
}
\description{
A combined dataset for Quality
}
\keyword{datasets}
