% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{quality}
\alias{quality}
\title{quality}
\format{
A data frame with 1037 rows and 12 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(quality)
}
\description{
A combined dataset for quality
}
\keyword{datasets}
