% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{state}
\alias{state}
\title{state}
\format{
A data frame with 100 rows and 11 columns covering 2012-2013.
}
\source{
Actuarial Data Master
}
\usage{
data(state)
}
\description{
A combined dataset for state
}
\keyword{datasets}
