% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{state}
\alias{state}
\title{state}
\format{
A data frame with 750 rows and 11 columns covering 2011-2025.

A data frame with 750 rows and 11 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(state)

data(state)
}
\description{
A combined dataset for state

A combined dataset for state
}
\keyword{datasets}
