% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{statement}
\alias{statement}
\title{statement}
\format{
A data frame with 181686 rows and 9 columns covering NA-NA.
}
\source{
Actuarial Data Master
}
\usage{
data(statement)
}
\description{
A combined dataset for statement
}
\keyword{datasets}
