% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{sub_class}
\alias{sub_class}
\title{sub_class}
\format{
A data frame with 80 rows and 9 columns covering 2011-2025.

A data frame with 80 rows and 9 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(sub_class)

data(sub_class)
}
\description{
A combined dataset for sub_class

A combined dataset for sub_class
}
\keyword{datasets}
