% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{subsidy_percent}
\alias{subsidy_percent}
\title{subsidy_percent}
\format{
A data frame with 9543 rows and 22 columns covering 2011-2025.

A data frame with 9543 rows and 22 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(subsidy_percent)

data(subsidy_percent)
}
\description{
A combined dataset for subsidy_percent

A combined dataset for subsidy_percent
}
\keyword{datasets}
