% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{SubsidyPercent}
\alias{SubsidyPercent}
\title{Actuarial Data Master Helper Datasets}
\format{
A data frame with 1331 rows and 17 columns covering 2011-2014.
}
\source{
Actuarial Data Master
}
\usage{
data(SubsidyPercent)
}
\description{
A combined dataset for SubsidyPercent
}
\keyword{datasets}
