% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{Type}
\alias{Type}
\title{Type}
\format{
A data frame with 8579 rows and 10 columns covering 2011-2025.
}
\source{
Actuarial Data Master
}
\usage{
data(Type)
}
\description{
A combined dataset for Type
}
\keyword{datasets}
