% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R, R/helper_data_2025-04-08.R
\docType{data}
\name{unit_discount}
\alias{unit_discount}
\title{unit_discount}
\format{
A data frame with 49741 rows and 28 columns covering 2011-2025.

A data frame with 49741 rows and 28 columns covering 2011-2025.
}
\source{
Actuarial Data Master

Actuarial Data Master
}
\usage{
data(unit_discount)

data(unit_discount)
}
\description{
A combined dataset for unit_discount

A combined dataset for unit_discount
}
\keyword{datasets}
