% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{UnitDiscount}
\alias{UnitDiscount}
\title{UnitDiscount}
\format{
A data frame with 11755 rows and 28 columns covering 2011-2014.
}
\source{
Actuarial Data Master
}
\usage{
data(UnitDiscount)
}
\description{
A combined dataset for UnitDiscount
}
\keyword{datasets}
